requirements.txt
flask
supabase
beautifulsoup4
requests
python-dotenv